fixed IssueTable
